commit in bulk! dont try this in your house!

+++ src/components/CarsTable.js
@@ -3,6 +3,7 @@ import { useEffect } from 'react';
 import CarService from './../services/CarService';
 import { DataGrid } from '@mui/x-data-grid';
 import { Button } from '@mui/material';
+import { Link } from 'react-router-dom';
 
 export default function CarsTable() {
   const [cars, setCars] = useState(null);
@@ -27,25 +28,41 @@ export default function CarsTable() {
     { field: 'licensePlate', headerName: 'Patente', width: 130 },
     { field: 'color', headerName: 'Color', width: 130 },
     {
-      field: 'owner',
-      headerName: 'Propietario',
+      field: 'owners',
+      headerName: 'Propietarios',
       width: 130,
       renderCell: (params) => {
         return (
-          <span>{`${params?.value?.firstname} ${params?.value?.lastname}`}</span>
+          <span>{`${params?.value[0]?.firstname} ${params?.value[0]?.lastname}`}</span>
         );
       },
     },
     {
       field: 'actions',
       headerName: 'Acciones',
-      width: 130,
+      width: 260,
       renderCell: (params) => {
-        const onClick = (e) => {
+        const handlerDelete = (e) => {
           e.preventDefault();
-          deleteCar(params.id);
+          if (window.confirm('¿Estas seguro?')) {
+            deleteCar(params.id);
+          }
         };
-        return <Button onClick={onClick}>Eliminar</Button>;
+        const editButton = (
+          <Button
+            to={'/cars/' + params.id}
+            variant='contained'
+            component={Link}
+          >
+            Editar
+          </Button>
+        );
+        const deleteButton = <Button onClick={handlerDelete}>Eliminar</Button>;
+        return (
+          <>
+            {editButton} {deleteButton}
+          </>
+        );
       },
     },
   ];
